+++ stlearn/tools/microenv/cci/permutation.py
@@ -13,6 +13,7 @@ from .base import lr, calc_neighbours, get_spot_lrs, get_lrs_scores, get_scores
 from .merge import merge
 from .perm_utils import get_lr_features, get_lr_bg
 
+
 # Newest method #
 def perform_spot_testing(
     adata: AnnData,
@@ -74,7 +75,7 @@ def perform_spot_testing(
     ######## Background per LR, but only for spots where LR has a score ########
     # Determine the indices of the spots where each LR has a score #
     cols = ["n_spots", "n_spots_sig", "n_spots_sig_pval"]
-    lr_summary = np.zeros((lr_scores.shape[1], 3), np.int)
+    lr_summary = np.zeros((lr_scores.shape[1], 3), np.int32)
     pvals = np.ones(lr_scores.shape, dtype=np.float64)
     pvals_adj = np.ones(lr_scores.shape, dtype=np.float64)
     log10pvals_adj = np.zeros(lr_scores.shape, dtype=np.float64)
@@ -349,7 +350,6 @@ def permutation(
     background: np.array = None,
     **kwargs,
 ) -> AnnData:
-
     """Permutation test for merged result
     Parameters
     ----------
